feat: comprehensive project cleanup and reorganization
- Move development docs to docs/development/ (8 files)
- Move user guides to docs/guides/ (4 files)
- Move utility scripts to scripts/ (2 files)
- Move test files to tests/ (3 files)
- Remove temporary files (exported_code.py, chart_summary.md)
- Remove duplicate files (demo_qc_charts.py from root)
- Remove empty directories (results/, configuration_exports/, data/)
- Create organized directory structure for better navigation
- Add cleanup documentation and summary

This creates a much cleaner, more professional project structure
with better organization and easier navigation.
