+++ CLEANUP_PLAN.md
@@ -1,0 +1,120 @@
+# PySEE Project Cleanup Plan
+
+## 🎯 **Cleanup Objectives**
+- Remove unnecessary files and duplicates
+- Organize documentation better
+- Clean up temporary and generated files
+- Consolidate similar functionality
+- Improve project structure
+
+## 📁 **Files to Remove**
+
+### **1. Temporary/Generated Files**
+- `exported_code.py` - Generated by example.py
+- `chart_summary.md` - Temporary analysis file
+- `__pycache__/` directories - Python cache files
+- `dist/` - Build artifacts (should be in .gitignore)
+- `pysee_bio.egg-info/` - Old package info
+- `pysee.egg-info/` - Package info (should be in .gitignore)
+
+### **2. Duplicate Files**
+- `demo_qc_charts.py` (root) - Duplicate of `examples/demo_qc_charts.py`
+- `examples/demo_qc_charts.py` - Keep this one, remove root version
+
+### **3. Development/Planning Documents (Move to docs/)**
+- `CUPY_TROUBLESHOOTING_PLAN.md` → `docs/development/`
+- `DOTPLOT_GUIDE.md` → `docs/guides/`
+- `DOTPLOT_QUICK_REFERENCE.md` → `docs/guides/`
+- `FEATURE_DELIVERY_ANALYSIS.md` → `docs/development/`
+- `GIT_WORKFLOW.md` → `docs/development/`
+- `GPU_ACCELERATION_ANALYSIS.md` → `docs/development/`
+- `IMMEDIATE_ACTION_PLAN.md` → `docs/development/`
+- `JUPYTER_INTEGRATION_PLAN.md` → `docs/development/`
+- `PERFORMANCE_TESTING_PLAN.md` → `docs/development/`
+- `SYSTEM_REQUIREMENTS_SIMPLIFIED.md` → `docs/guides/`
+- `VERSION_STRATEGY.md` → `docs/development/`
+
+### **4. Utility Scripts (Move to scripts/)**
+- `check_system_requirements.py` → `scripts/`
+- `decide_gpu_approach.py` → `scripts/`
+- `generate_readme_examples.py` → `scripts/`
+- `run_performance_tests.py` → `scripts/` (or remove if duplicate)
+
+### **5. Test Files (Move to tests/)**
+- `test_gpu_simple.py` → `tests/`
+- `test_heatmap.py` → `tests/`
+- `test_large_datasets_16gb.py` → `tests/`
+- `test_panel_configuration.py` → `tests/`
+- `test_publication_export.py` → `tests/`
+- `test_qc.py` → `tests/`
+
+### **6. Empty Directories**
+- `results/` - Empty, remove
+- `configuration_exports/` - Empty, remove
+- `data/` - Empty, remove
+
+## 📂 **New Directory Structure**
+
+```
+docs/
+├── development/
+│   ├── CUPY_TROUBLESHOOTING_PLAN.md
+│   ├── FEATURE_DELIVERY_ANALYSIS.md
+│   ├── GIT_WORKFLOW.md
+│   ├── GPU_ACCELERATION_ANALYSIS.md
+│   ├── IMMEDIATE_ACTION_PLAN.md
+│   ├── JUPYTER_INTEGRATION_PLAN.md
+│   ├── PERFORMANCE_TESTING_PLAN.md
+│   └── VERSION_STRATEGY.md
+├── guides/
+│   ├── DOTPLOT_GUIDE.md
+│   ├── DOTPLOT_QUICK_REFERENCE.md
+│   └── SYSTEM_REQUIREMENTS_SIMPLIFIED.md
+└── setup/
+    └── (existing files)
+
+scripts/
+├── check_system_requirements.py
+├── decide_gpu_approach.py
+├── generate_readme_examples.py
+├── run_performance_tests.py
+└── (existing migration scripts)
+
+tests/
+├── test_gpu_simple.py
+├── test_heatmap.py
+├── test_large_datasets_16gb.py
+├── test_panel_configuration.py
+├── test_publication_export.py
+├── test_qc.py
+└── (existing test structure)
+```
+
+## 🧹 **Cleanup Actions**
+
+### **Phase 1: Remove Temporary Files**
+1. Delete generated files
+2. Remove empty directories
+3. Clean up cache files
+
+### **Phase 2: Reorganize Documentation**
+1. Create new directory structure
+2. Move planning documents to `docs/development/`
+3. Move guides to `docs/guides/`
+
+### **Phase 3: Reorganize Scripts and Tests**
+1. Move utility scripts to `scripts/`
+2. Move test files to `tests/`
+3. Remove duplicates
+
+### **Phase 4: Update References**
+1. Update import paths
+2. Update documentation links
+3. Update CI workflows if needed
+
+## ✅ **Benefits**
+- Cleaner project structure
+- Better organization
+- Easier navigation
+- Reduced clutter
+- Professional appearance
