--- DOTPLOT_QUICK_REFERENCE.md
@@ -1,56 +1,0 @@
-# DotPlot Quick Reference Card
-
-## 🔵 Circle Size = % of Cells Expressing Gene
-- **Big Circle** = Many cells express this gene
-- **Small Circle** = Few cells express this gene
-
-## 🎨 Color = Expression Level
-- **Red** = High expression
-- **Yellow** = Medium expression
-- **Blue** = Low expression
-
-## 📊 Reading the Chart
-
-| Element | Meaning |
-|---------|---------|
-| **Row** | Gene name |
-| **Column** | Cell group/cluster |
-| **Dot Position** | Gene in specific group |
-| **Dot Size** | How many cells express it |
-| **Dot Color** | How much they express it |
-
-## 🔍 Quick Questions
-
-### **"Is this gene specific to this cell type?"**
-- **Big red dot in one group only** = YES, specific marker
-- **Small dots everywhere** = NO, not a marker
-
-### **"How pure is this cell cluster?"**
-- **Clear marker patterns** = Good purity
-- **Mixed patterns** = Mixed population
-
-### **"Which genes are most important?"**
-- **Biggest, reddest dots** = Most important markers
-
-## 🧬 Common Patterns
-
-### **Cell Type Markers**
-- **B cells**: MS4A1, CD19 (big red dots)
-- **NK cells**: NKG7, GNLY (big red dots)
-- **T cells**: CD3D, CD3E (big red dots)
-
-### **Quality Indicators**
-- **Clear separation** = Good clustering
-- **Overlapping patterns** = Need better separation
-
-## ⚡ Quick Tips
-
-1. **Hover** for exact numbers
-2. **Look for patterns**, not individual dots
-3. **Compare sizes** within the same gene
-4. **Use known markers** to validate clusters
-5. **Check both size AND color**
-
----
-
-**Remember: Big + Red = Strong marker for that cell type! 🎯**
