--- FEATURE_DELIVERY_ANALYSIS.md
@@ -1,228 +1,0 @@
-# PySEE Feature Delivery Analysis & Priority Plan
-
-## 🎯 **Current Project Status**
-
-### **✅ COMPLETED FEATURES (Ready for Release)**
-1. **Core Architecture** ✅
-   - AnnData integration (`pysee/core/data.py`)
-   - Dashboard engine (`pysee/core/dashboard.py`)
-   - Panel management system
-
-2. **Visualization Panels** ✅
-   - **UMAP Panel** (`pysee/panels/umap.py`) - Interactive scatter plots
-   - **Violin Panel** (`pysee/panels/violin.py`) - Gene expression violin plots
-   - **Heatmap Panel** (`pysee/panels/heatmap.py`) - Gene expression heatmaps
-   - **QC Panel** (`pysee/panels/qc.py`) - Quality control metrics
-
-3. **Development Infrastructure** ✅
-   - Git workflow (main/develop/feature branches)
-   - CI/CD pipeline (GitHub Actions)
-   - Testing framework
-   - Documentation (README, examples)
-
-### **📊 CURRENT BRANCH STATUS**
-- **Current Branch**: `feature/qc-metrics-panel`
-- **Status**: QC Panel completed, ready to merge
-- **Next**: Merge to develop → Create v0.1.2 release
-
----
-
-## 🚀 **PRIORITY ACTION PLAN**
-
-### **PHASE 1: IMMEDIATE (This Week) - Release v0.1.2**
-
-#### **1.1 Complete Current Feature Branch** 🔥 **HIGH PRIORITY**
-```bash
-# Current status: QC panel completed on feature/qc-metrics-panel
-git add .
-git commit -m "feat: complete QC metrics panel implementation"
-git push origin feature/qc-metrics-panel
-```
-
-#### **1.2 Merge to Develop** 🔥 **HIGH PRIORITY**
-```bash
-# Create PR and merge QC panel
-# Then merge develop to main
-git checkout develop
-git merge feature/qc-metrics-panel
-git push origin develop
-```
-
-#### **1.3 Create v0.1.2 Release** 🔥 **HIGH PRIORITY**
-```bash
-# Tag and release v0.1.2
-git tag v0.1.2
-git push origin v0.1.2
-# Create GitHub release with changelog
-```
-
-**Why This Priority**:
-- 4 complete panels ready for users
-- Solid foundation for further development
-- Get user feedback early
-
-### **PHASE 2: SHORT-TERM (Next 2 Weeks) - Core Features**
-
-#### **2.1 DotPlot Panel** 📊 **HIGH PRIORITY**
-- **Effort**: 2-3 days
-- **Impact**: High - most requested feature after heatmap
-- **Description**: Marker gene visualization with dot plots
-- **Implementation**: `pysee/panels/dotplot.py`
-
-#### **2.2 Advanced Selection Tools** 🎯 **MEDIUM PRIORITY**
-- **Effort**: 1-2 days
-- **Impact**: Medium - improves user experience
-- **Description**: Lasso selection, polygon selection
-- **Implementation**: Enhance existing panels
-
-#### **2.3 Jupyter Widget Integration** 📓 **MEDIUM PRIORITY**
-- **Effort**: 2-3 days
-- **Impact**: Medium - better notebook experience
-- **Description**: Native Jupyter widgets for PySEE
-- **Implementation**: `pysee/widgets/`
-
-### **PHASE 3: MEDIUM-TERM (Next Month) - Advanced Features**
-
-#### **3.1 Multi-Modal Data Support** 🔄 **MEDIUM PRIORITY**
-- **Effort**: 3-4 days
-- **Impact**: Medium - supports more data types
-- **Description**: CITE-seq, ATAC-seq, multi-omics
-- **Implementation**: Extend data handling
-
-#### **3.2 Export & Sharing** 💾 **LOW PRIORITY**
-- **Effort**: 1-2 days
-- **Impact**: Low - nice to have
-- **Description**: Export plots, share dashboards
-- **Implementation**: Add export methods
-
-#### **3.3 Performance Optimization** ⚡ **LOW PRIORITY**
-- **Effort**: 2-3 days
-- **Impact**: Low - optimization
-- **Description**: GPU acceleration, memory optimization
-- **Implementation**: Based on performance analysis
-
----
-
-## 📋 **DETAILED TODO LIST (Prioritized)**
-
-### **🔥 CRITICAL (This Week)**
-1. **Complete QC Panel Branch**
-   - [ ] Commit current changes
-   - [ ] Push to GitHub
-   - [ ] Create PR
-   - [ ] Merge to develop
-
-2. **Create v0.1.2 Release**
-   - [ ] Update version numbers
-   - [ ] Create changelog
-   - [ ] Tag release
-   - [ ] Publish to PyPI
-
-3. **Update Documentation**
-   - [ ] Update README with 4 panels
-   - [ ] Create quickstart guide
-   - [ ] Add examples for all panels
-
-### **📊 HIGH PRIORITY (Next 2 Weeks)**
-4. **DotPlot Panel Implementation**
-   - [ ] Create `pysee/panels/dotplot.py`
-   - [ ] Implement marker gene visualization
-   - [ ] Add interactive features
-   - [ ] Write tests
-
-5. **Advanced Selection Tools**
-   - [ ] Add lasso selection to UMAP panel
-   - [ ] Add polygon selection
-   - [ ] Improve selection UX
-
-6. **Jupyter Widget Integration**
-   - [ ] Create widget base class
-   - [ ] Convert panels to widgets
-   - [ ] Add notebook examples
-
-### **🔄 MEDIUM PRIORITY (Next Month)**
-7. **Multi-Modal Data Support**
-   - [ ] Extend data handling for CITE-seq
-   - [ ] Add ATAC-seq support
-   - [ ] Create multi-omics examples
-
-8. **Export & Sharing Features**
-   - [ ] Add plot export methods
-   - [ ] Create dashboard sharing
-   - [ ] Add PDF/PNG export
-
-### **⚡ LOW PRIORITY (Future)**
-9. **Performance Optimization**
-   - [ ] GPU acceleration (when CUDA is fixed)
-   - [ ] Memory optimization
-   - [ ] Large dataset handling
-
-10. **Cloud Deployment**
-    - [ ] Google Colab examples
-    - [ ] AWS deployment guides
-    - [ ] Docker containers
-
----
-
-## 🎯 **RECOMMENDED IMMEDIATE ACTIONS**
-
-### **Today:**
-1. **Complete current branch** - Commit and push QC panel changes
-2. **Create PR** - Merge QC panel to develop
-3. **Update documentation** - Reflect 4 complete panels
-
-### **This Week:**
-1. **Release v0.1.2** - Tag and publish current features
-2. **Start DotPlot panel** - Begin next major feature
-3. **Gather user feedback** - Share v0.1.2 with community
-
-### **Next Week:**
-1. **Complete DotPlot panel** - Finish marker gene visualization
-2. **Add selection tools** - Improve user interaction
-3. **Plan v0.2 features** - Based on user feedback
-
----
-
-## 💡 **KEY INSIGHTS**
-
-### **What's Working Well:**
-- ✅ Solid core architecture
-- ✅ 4 complete, functional panels
-- ✅ Good development workflow
-- ✅ Comprehensive testing framework
-
-### **What Needs Focus:**
-- 🎯 **Feature delivery** over performance optimization
-- 🎯 **User experience** over technical complexity
-- 🎯 **Release early and often** for feedback
-
-### **What to Defer:**
-- ⏳ GPU acceleration (complex, not critical)
-- ⏳ Performance optimization (premature)
-- ⏳ Cloud deployment (can be added later)
-
----
-
-## 🚀 **SUCCESS METRICS**
-
-### **v0.1.2 Release Success:**
-- [ ] 4 functional panels
-- [ ] Clean documentation
-- [ ] Working examples
-- [ ] User feedback
-
-### **v0.2 Release Success:**
-- [ ] 5+ panels (add DotPlot)
-- [ ] Advanced selection tools
-- [ ] Jupyter integration
-- [ ] Community adoption
-
-### **Long-term Success:**
-- [ ] 10+ panels
-- [ ] Multi-modal support
-- [ ] Performance optimization
-- [ ] Cloud deployment
-
----
-
-**Bottom Line**: Focus on feature delivery and user experience. Performance optimization can wait until we have more users and feedback! 🎯
