+++ docs/development/CUPY_TROUBLESHOOTING_PLAN.md
@@ -1,0 +1,154 @@
+# CuPy Troubleshooting Plan
+
+## 🔍 **Current Issue Analysis**
+
+**Problem**: CuPy is installed but GPU operations fail with:
+```
+CuPy failed to load nvrtc64_120_0.dll: FileNotFoundError: Could not find module 'nvrtc64_120_0.dll'
+```
+
+**Root Cause**: Missing CUDA runtime libraries on Windows system.
+
+## 🎯 **Solution Options (Ranked by Difficulty)**
+
+### **Option 1: Install CUDA Toolkit (Recommended)**
+**Difficulty**: Medium | **Time**: 30-60 minutes | **Success Rate**: High
+
+**Steps**:
+1. **Download CUDA Toolkit 12.x**:
+   - Go to: https://developer.nvidia.com/cuda-downloads
+   - Select: Windows → x86_64 → 11 → exe (local)
+   - Download: `cuda_12.x.x_windows.exe` (~3-4 GB)
+
+2. **Install CUDA Toolkit**:
+   ```bash
+   # Run the installer as administrator
+   # Choose "Custom" installation
+   # Select: CUDA Toolkit, CUDA Samples, CUDA Documentation
+   # Install to default location: C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.x
+   ```
+
+3. **Set Environment Variables**:
+   ```bash
+   # Add to system PATH:
+   C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.x\bin
+   C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.x\libnvvp
+
+   # Set CUDA_PATH:
+   CUDA_PATH=C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.x
+   ```
+
+4. **Restart and Test**:
+   ```bash
+   python test_gpu_simple.py
+   ```
+
+### **Option 2: Use Conda (Alternative)**
+**Difficulty**: Easy | **Time**: 15-30 minutes | **Success Rate**: Medium
+
+**Steps**:
+1. **Install Miniconda** (if not already installed):
+   ```bash
+   # Download from: https://docs.conda.io/en/latest/miniconda.html
+   # Install with default settings
+   ```
+
+2. **Create new environment with CUDA**:
+   ```bash
+   conda create -n pysee-gpu python=3.11
+   conda activate pysee-gpu
+   conda install -c conda-forge cupy cudatoolkit=12.0
+   ```
+
+3. **Test**:
+   ```bash
+   python test_gpu_simple.py
+   ```
+
+### **Option 3: Use Docker (Advanced)**
+**Difficulty**: Hard | **Time**: 60+ minutes | **Success Rate**: High
+
+**Steps**:
+1. **Install Docker Desktop**:
+   - Download from: https://www.docker.com/products/docker-desktop
+   - Enable WSL2 backend
+
+2. **Use NVIDIA CUDA Docker image**:
+   ```bash
+   docker run --gpus all -it nvidia/cuda:12.0-devel-ubuntu20.04
+   ```
+
+3. **Install PySEE in container**:
+   ```bash
+   pip install cupy-cuda12x pysee
+   ```
+
+### **Option 4: Skip GPU for Now (Pragmatic)**
+**Difficulty**: None | **Time**: 0 minutes | **Success Rate**: 100%
+
+**Decision**: Focus on CPU optimization and cloud deployment instead of local GPU acceleration.
+
+## 🎯 **Recommended Approach**
+
+### **For PySEE Development (Recommended)**:
+**Skip GPU for now** and focus on:
+1. **CPU optimization** for PySEE
+2. **Cloud deployment** for large datasets
+3. **WebGL acceleration** for visualization
+4. **Smart sampling** for large datasets
+
+**Why**:
+- GPU acceleration adds complexity
+- Most users will use cloud for large datasets anyway
+- CPU optimization is more universally applicable
+- WebGL provides browser GPU acceleration
+
+### **For Future GPU Support**:
+**Option 1 (CUDA Toolkit)** when you have time:
+- Most comprehensive solution
+- Enables full GPU acceleration
+- Good for future development
+
+## 📋 **Implementation Plan**
+
+### **Immediate (Today)**:
+1. **Update PySEE to work without GPU**:
+   - Remove GPU dependencies from core code
+   - Add graceful fallback for GPU operations
+   - Focus on CPU optimization
+
+2. **Update documentation**:
+   - Clarify that GPU is optional
+   - Provide cloud deployment examples
+   - Document CPU-only usage
+
+### **Short-term (This Week)**:
+1. **Implement CUDA Toolkit installation** (Option 1)
+2. **Test GPU acceleration** with proper CUDA setup
+3. **Add GPU support** to PySEE if successful
+
+### **Long-term (Future)**:
+1. **Hybrid CPU/GPU architecture**
+2. **Cloud GPU deployment**
+3. **RAPIDS integration**
+
+## 🚀 **Quick Decision Matrix**
+
+| Option | Time | Complexity | Success Rate | Recommendation |
+|--------|------|------------|--------------|----------------|
+| Skip GPU | 0 min | None | 100% | ✅ **Do this now** |
+| Conda | 30 min | Low | 70% | ⚠️ Try if you want GPU |
+| CUDA Toolkit | 60 min | Medium | 90% | 🔄 Do later |
+| Docker | 120 min | High | 95% | ❌ Overkill |
+
+## 💡 **My Recommendation**
+
+**For PySEE v0.1.2**: Skip GPU acceleration and focus on:
+1. **CPU optimization**
+2. **Cloud deployment examples**
+3. **WebGL acceleration**
+4. **Smart sampling strategies**
+
+**For PySEE v0.2+**: Implement CUDA Toolkit (Option 1) when you have time.
+
+**Why**: GPU acceleration is nice-to-have, but CPU optimization and cloud deployment are more important for user experience.
