+++ docs/development/IMMEDIATE_ACTION_PLAN.md
@@ -1,0 +1,178 @@
+# PySEE Immediate Action Plan
+
+## 🎯 **Current Status Summary**
+
+### **✅ What's Done:**
+- **4 Complete Panels**: UMAP, Violin, Heatmap, QC
+- **Core Architecture**: Solid foundation
+- **Development Infrastructure**: CI/CD, testing, documentation
+- **Performance Analysis**: GPU/CPU analysis completed (for planning)
+
+### **🔄 What's In Progress:**
+- **Current Branch**: `feature/qc-metrics-panel` (QC panel completed)
+- **Ready to Merge**: QC panel implementation
+
+### **📋 What's Next:**
+- **Immediate**: Complete current branch and create v0.1.2 release
+- **Short-term**: Add DotPlot panel and advanced selection tools
+
+---
+
+## 🚀 **IMMEDIATE ACTIONS (Today)**
+
+### **Step 1: Complete Current Branch** ⏱️ 15 minutes
+```bash
+# 1. Add all changes
+git add .
+
+# 2. Commit with descriptive message
+git commit -m "feat: complete QC metrics panel and performance analysis
+
+- Complete QC panel implementation with all features
+- Add comprehensive performance testing framework
+- Add system requirements checking and user guidance
+- Add cloud deployment examples and GPU analysis
+- Update documentation and master todos
+- Ready for v0.1.2 release with 4 complete panels"
+
+# 3. Push to GitHub
+git push origin feature/qc-metrics-panel
+```
+
+### **Step 2: Create Pull Request** ⏱️ 5 minutes
+1. Go to GitHub: https://github.com/Linnnnberg/PySEE
+2. Create PR: `feature/qc-metrics-panel` → `develop`
+3. Title: "feat: Complete QC Metrics Panel and Performance Framework"
+4. Description: List all completed features
+
+### **Step 3: Merge to Develop** ⏱️ 5 minutes
+```bash
+# After PR is approved and merged
+git checkout develop
+git pull origin develop
+```
+
+---
+
+## 📊 **THIS WEEK'S GOALS**
+
+### **Day 1-2: Release v0.1.2**
+- [ ] Merge QC panel to develop
+- [ ] Update version numbers
+- [ ] Create changelog
+- [ ] Tag and release v0.1.2
+- [ ] Publish to PyPI
+
+### **Day 3-5: Start Next Features**
+- [ ] Begin DotPlot panel implementation
+- [ ] Plan advanced selection tools
+- [ ] Update documentation
+
+---
+
+## 🎯 **PRIORITY FEATURES (Next 2 Weeks)**
+
+### **1. DotPlot Panel** 🔥 **HIGH PRIORITY**
+**Why**: Most requested feature after heatmap
+**Effort**: 2-3 days
+**Impact**: High - marker gene visualization
+
+**Implementation Plan**:
+```python
+# pysee/panels/dotplot.py
+class DotPlotPanel(BasePanel):
+    def __init__(self, panel_id, genes, groupby, ...):
+        # Marker gene dot plot visualization
+        # Interactive features
+        # Color/size mapping
+```
+
+### **2. Advanced Selection Tools** 🎯 **MEDIUM PRIORITY**
+**Why**: Improves user experience
+**Effort**: 1-2 days
+**Impact**: Medium - better interaction
+
+**Implementation Plan**:
+- Add lasso selection to UMAP panel
+- Add polygon selection
+- Improve selection UX
+
+### **3. Jupyter Widget Integration** 📓 **MEDIUM PRIORITY**
+**Why**: Better notebook experience
+**Effort**: 2-3 days
+**Impact**: Medium - notebook users
+
+**Implementation Plan**:
+- Create widget base class
+- Convert panels to widgets
+- Add notebook examples
+
+---
+
+## 📋 **DEFERRED ITEMS (Future)**
+
+### **Performance Optimization** ⏳
+- GPU acceleration (when CUDA is fixed)
+- Memory optimization
+- Large dataset handling
+
+### **Cloud Deployment** ⏳
+- Google Colab examples
+- AWS deployment guides
+- Docker containers
+
+### **Advanced Features** ⏳
+- Multi-modal data support
+- Export/sharing features
+- Custom panel types
+
+---
+
+## 💡 **KEY DECISIONS MADE**
+
+### **✅ Focus on Feature Delivery**
+- Prioritize user-facing features over performance optimization
+- Get v0.1.2 released quickly for user feedback
+- Build on solid foundation we have
+
+### **✅ Defer Performance Optimization**
+- GPU acceleration is complex and not critical
+- Performance analysis is done for future planning
+- Focus on user experience first
+
+### **✅ Cloud-First for Large Datasets**
+- Don't struggle with large datasets locally
+- Use cloud for 100K+ cell datasets
+- Keep local usage simple and fast
+
+---
+
+## 🎯 **SUCCESS METRICS**
+
+### **v0.1.2 Release (This Week)**
+- [ ] 4 functional panels
+- [ ] Clean documentation
+- [ ] Working examples
+- [ ] User feedback
+
+### **v0.2 Release (Next Month)**
+- [ ] 5+ panels (add DotPlot)
+- [ ] Advanced selection tools
+- [ ] Jupyter integration
+- [ ] Community adoption
+
+---
+
+## 🚀 **READY TO EXECUTE**
+
+**Current Status**: Ready to complete QC panel branch and create v0.1.2 release
+
+**Next Action**: Run the git commands above to complete current branch
+
+**Timeline**: v0.1.2 release this week, v0.2 features next month
+
+**Focus**: Feature delivery and user experience over performance optimization
+
+---
+
+**Let's ship v0.1.2 and get user feedback! 🚀**
