+++ docs/guides/SYSTEM_REQUIREMENTS_SIMPLIFIED.md
@@ -1,0 +1,92 @@
+# PySEE System Requirements - Simplified Approach
+
+## 🎯 Realistic System Requirements
+
+### **Local Development (Recommended)**
+- **Minimum**: 8 GB RAM
+- **Recommended**: 16 GB RAM
+- **Optimal**: 32 GB RAM
+- **Use for**: Development, testing, small-medium datasets (up to 68K cells)
+
+### **Cloud/Server (For Large Datasets)**
+- **Google Colab**: Free tier (12 GB RAM) - good for medium datasets
+- **Google Colab Pro**: 25 GB RAM - good for large datasets
+- **AWS/GCP**: 32+ GB RAM - optimal for very large datasets
+- **Use for**: Large datasets (100K+ cells), production analysis
+
+## 📊 Dataset Size Guidelines
+
+| Dataset Size | Cells | Memory | Local (16GB) | Cloud/Server |
+|--------------|-------|--------|--------------|--------------|
+| Small        | 3K    | 350 MB | ✅ Perfect   | ✅ Perfect   |
+| Medium       | 68K   | 8.5 GB | ⚠️ Caution   | ✅ Perfect   |
+| Large        | 100K+ | 15+ GB | ❌ Not recommended | ✅ Recommended |
+| Very Large   | 1M+   | 100+ GB| ❌ Not feasible | ✅ Required |
+
+## 🚀 Recommended Workflow
+
+### **1. Local Development**
+```python
+# Perfect for local development
+import scanpy as sc
+from pysee import PySEE
+
+# Load small dataset
+adata = sc.datasets.pbmc3k()  # 3K cells, 350 MB
+app = PySEE(adata)
+# ... add panels and test
+```
+
+### **2. Cloud for Large Datasets**
+```python
+# Google Colab example
+!pip install pysee scanpy
+
+import scanpy as sc
+from pysee import PySEE
+
+# Load large dataset (works in cloud)
+adata = sc.datasets.pbmc68k_reduced()  # 68K cells, 8.5 GB
+app = PySEE(adata)
+# ... full analysis
+```
+
+### **3. Server Deployment**
+```python
+# For production or very large datasets
+# Deploy on AWS/GCP with 32+ GB RAM
+# Use backed mode for memory efficiency
+adata = ad.read_h5ad('large_dataset.h5ad', backed='r')
+app = PySEE(adata)
+```
+
+## 💡 User Guidance
+
+### **For Researchers with 16 GB RAM:**
+1. **Use locally**: Small datasets (3K cells) for development
+2. **Use cloud**: Medium-large datasets (68K+ cells) for analysis
+3. **Don't try**: Very large datasets (1M+ cells) locally
+
+### **For Researchers with 32+ GB RAM:**
+1. **Use locally**: Small-medium datasets (up to 68K cells)
+2. **Use cloud**: Large datasets (100K+ cells) for better performance
+3. **Consider local**: Very large datasets with backed mode
+
+### **For Production/Teams:**
+1. **Deploy on cloud**: Always use cloud for large datasets
+2. **Use HPC**: For very large datasets (1M+ cells)
+3. **Local development**: Only for small datasets and testing
+
+## 🎯 Key Takeaways
+
+1. **Don't overcomplicate**: Most users should use cloud for large datasets
+2. **Focus on cloud deployment**: Make PySEE work great in cloud environments
+3. **Simple local usage**: Keep local usage simple and realistic
+4. **Clear boundaries**: 16 GB = up to 68K cells, 32+ GB = up to 100K cells, Cloud = 100K+ cells
+
+## 🚀 Next Steps
+
+1. **Simplify system requirements**: Remove complex memory strategies
+2. **Focus on cloud deployment**: Add Google Colab, AWS examples
+3. **Clear user guidance**: Simple rules for when to use local vs cloud
+4. **Remove memory hacks**: Don't encourage users to struggle with limited RAM
