+++ scripts/decide_gpu_approach.py
@@ -1,0 +1,164 @@
+"""
+Decision helper for GPU approach in PySEE.
+
+This script helps you decide whether to pursue GPU acceleration
+or focus on CPU optimization and cloud deployment.
+"""
+
+import sys
+from pathlib import Path
+
+# Add project root to path
+project_root = Path(__file__).parent
+sys.path.insert(0, str(project_root))
+
+
+def analyze_current_situation():
+    """Analyze current GPU situation."""
+    print("🔍 Current GPU Situation Analysis")
+    print("=" * 50)
+
+    # Check system resources
+    import psutil
+
+    system_info = {
+        "cpu_count": psutil.cpu_count(),
+        "memory_gb": psutil.virtual_memory().total / (1024**3),
+    }
+
+    print(f"System Resources:")
+    print(f"  CPU Cores: {system_info['cpu_count']}")
+    print(f"  Memory: {system_info['memory_gb']:.1f} GB")
+
+    # Check GPU status
+    gpu_status = check_gpu_status()
+    print(f"  GPU Status: {gpu_status}")
+
+    return system_info, gpu_status
+
+
+def check_gpu_status():
+    """Check GPU status."""
+    try:
+        import cupy as cp
+
+        if cp.cuda.is_available():
+            try:
+                # Test if GPU operations work
+                test_array = cp.array([1, 2, 3])
+                result = test_array * 2
+                return "✅ GPU operations working"
+            except Exception as e:
+                return f"⚠️ GPU operations failing: {str(e)[:50]}..."
+        else:
+            return "❌ CUDA not available"
+    except ImportError:
+        return "❌ CuPy not installed"
+
+
+def provide_recommendations(system_info, gpu_status):
+    """Provide recommendations based on current situation."""
+    print("\n💡 Recommendations")
+    print("=" * 50)
+
+    # Analyze situation
+    has_working_gpu = "working" in gpu_status
+    has_sufficient_ram = system_info["memory_gb"] >= 16
+
+    print(f"Analysis:")
+    print(f"  Working GPU: {has_working_gpu}")
+    print(f"  Sufficient RAM: {has_sufficient_ram} ({system_info['memory_gb']:.1f} GB)")
+
+    # Provide recommendations
+    if has_working_gpu and has_sufficient_ram:
+        print(f"\n🎯 Recommendation: **Hybrid Approach**")
+        print(f"  - Use GPU for heavy computations")
+        print(f"  - Use CPU for visualization")
+        print(f"  - Implement both CPU and GPU code paths")
+        print(f"  - Time investment: 2-4 hours")
+
+    elif has_sufficient_ram:
+        print(f"\n🎯 Recommendation: **CPU-First Approach**")
+        print(f"  - Focus on CPU optimization")
+        print(f"  - Use cloud for large datasets")
+        print(f"  - Implement WebGL acceleration")
+        print(f"  - Time investment: 1-2 hours")
+
+    else:
+        print(f"\n🎯 Recommendation: **Cloud-First Approach**")
+        print(f"  - Use local only for small datasets")
+        print(f"  - Use cloud for medium+ datasets")
+        print(f"  - Focus on cloud deployment examples")
+        print(f"  - Time investment: 1 hour")
+
+
+def show_implementation_options():
+    """Show implementation options."""
+    print(f"\n🚀 Implementation Options")
+    print("=" * 50)
+
+    print(f"1. **Skip GPU for Now (Recommended for v0.1.2)**")
+    print(f"   ✅ Pros: Fast, simple, works everywhere")
+    print(f"   ❌ Cons: No GPU acceleration")
+    print(f"   ⏱️ Time: 0 hours")
+    print(f"   🎯 Best for: Getting PySEE v0.1.2 released quickly")
+
+    print(f"\n2. **Fix GPU Setup (Option for v0.2+)**")
+    print(f"   ✅ Pros: GPU acceleration, future-proof")
+    print(f"   ❌ Cons: Complex, time-consuming")
+    print(f"   ⏱️ Time: 2-4 hours")
+    print(f"   🎯 Best for: Long-term development")
+
+    print(f"\n3. **Hybrid Approach (Best of Both)**")
+    print(f"   ✅ Pros: CPU fallback, GPU when available")
+    print(f"   ❌ Cons: More complex code")
+    print(f"   ⏱️ Time: 3-5 hours")
+    print(f"   🎯 Best for: Production-ready solution")
+
+
+def show_next_steps():
+    """Show next steps."""
+    print(f"\n📋 Next Steps")
+    print("=" * 50)
+
+    print(f"**Immediate (Today):**")
+    print(f"  1. Update PySEE to work without GPU")
+    print(f"  2. Focus on CPU optimization")
+    print(f"  3. Add cloud deployment examples")
+    print(f"  4. Implement WebGL acceleration")
+
+    print(f"\n**Short-term (This Week):**")
+    print(f"  1. Test PySEE with small-medium datasets")
+    print(f"  2. Create cloud deployment guides")
+    print(f"  3. Optimize CPU performance")
+
+    print(f"\n**Long-term (Future):**")
+    print(f"  1. Fix GPU setup when you have time")
+    print(f"  2. Add GPU acceleration to PySEE")
+    print(f"  3. Implement hybrid CPU/GPU architecture")
+
+
+def main():
+    """Main decision helper."""
+    print("🎯 PySEE GPU Decision Helper")
+    print("=" * 60)
+
+    # Analyze current situation
+    system_info, gpu_status = analyze_current_situation()
+
+    # Provide recommendations
+    provide_recommendations(system_info, gpu_status)
+
+    # Show implementation options
+    show_implementation_options()
+
+    # Show next steps
+    show_next_steps()
+
+    print(f"\n🎯 **My Recommendation:**")
+    print(f"Skip GPU for now and focus on CPU optimization + cloud deployment.")
+    print(f"This will get PySEE v0.1.2 released faster and serve more users.")
+
+
+if __name__ == "__main__":
+    main()
